+++ src/pipelayer/filter.py
@@ -2,27 +2,24 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from collections.abc import Iterable
-from functools import wraps
-from typing import Any, Callable, List, Optional, Tuple, Union, cast
+from typing import Any, List, Optional, Union
 
-from pipelayer.context import Context
-from pipelayer.enum import Action, State
 from pipelayer.event_args import FilterEventArgs
-from pipelayer.protocol import FilterEventHandlerT, IFilter
+from pipelayer.protocol import FilterEventHandlerT
 
 
-class EventHandlerList(List[FilterEventHandlerT]):
+class FilterEventHandlerList(List[FilterEventHandlerT]):
     def append(self, handler: FilterEventHandlerT) -> None:
         super().append(handler)
 
-    def __iadd__(self, handlers: Union[FilterEventHandlerT, Iterable[FilterEventHandlerT]]) -> EventHandlerList:
-        return EventHandlerList(super().__iadd__(handlers if isinstance(handlers, Iterable) else [handlers]))
+    def __iadd__(self, handlers: Union[FilterEventHandlerT, Iterable[FilterEventHandlerT]]) -> FilterEventHandlerList:
+        return FilterEventHandlerList(super().__iadd__(handlers if isinstance(handlers, Iterable) else [handlers]))
 
-    def __add__(self, handlers: Union[FilterEventHandlerT, Iterable[FilterEventHandlerT]]) -> EventHandlerList:
-        return EventHandlerList(super().__add__(
-            EventHandlerList(handlers)
+    def __add__(self, handlers: Union[FilterEventHandlerT, Iterable[FilterEventHandlerT]]) -> FilterEventHandlerList:
+        return FilterEventHandlerList(super().__add__(
+            FilterEventHandlerList(handlers)
             if isinstance(handlers, Iterable)
-            else EventHandlerList([handlers])
+            else FilterEventHandlerList([handlers])
         ))
 
 
@@ -37,9 +34,9 @@ class Filter(ABC):
     ) -> None:
         self.__name = name or self.__class__.__name__
 
-        self.__start: EventHandlerList = EventHandlerList()
-        self.__exit: EventHandlerList = EventHandlerList()
-        self.__end: EventHandlerList = EventHandlerList()
+        self.__start: FilterEventHandlerList = FilterEventHandlerList()
+        self.__exit: FilterEventHandlerList = FilterEventHandlerList()
+        self.__end: FilterEventHandlerList = FilterEventHandlerList()
 
     # region Properties
 
@@ -50,32 +47,32 @@ class Filter(ABC):
     # region Event Handlers
 
     @property
-    def start(self) -> EventHandlerList:
+    def start(self) -> FilterEventHandlerList:
         return self.__start
 
     @start.setter
-    def start(self, value: EventHandlerList) -> None:
-        if not isinstance(value, EventHandlerList):
+    def start(self, value: FilterEventHandlerList) -> None:
+        if not isinstance(value, FilterEventHandlerList):
             raise TypeError("Value must be an instance of EventHandlerList")
         self.__start = value
 
     @property
-    def exit(self) -> EventHandlerList:
+    def exit(self) -> FilterEventHandlerList:
         return self.__exit
 
     @exit.setter
-    def exit(self, value: EventHandlerList) -> None:
-        if not isinstance(value, EventHandlerList):
+    def exit(self, value: FilterEventHandlerList) -> None:
+        if not isinstance(value, FilterEventHandlerList):
             raise TypeError("Value must be an instance of EventHandlerList")
         self.__exit = value
 
     @property
-    def end(self) -> EventHandlerList:
+    def end(self) -> FilterEventHandlerList:
         return self.__end
 
     @end.setter
-    def end(self, value: EventHandlerList) -> None:
-        if not isinstance(value, EventHandlerList):
+    def end(self, value: FilterEventHandlerList) -> None:
+        if not isinstance(value, FilterEventHandlerList):
             raise TypeError("Value must be an instance of EventHandlerList")
         self.__end = value
 
@@ -101,59 +98,3 @@ class Filter(ABC):
         [e(self, args) for e in self.end]
 
     # endregion
-
-
-def _parse_args(*args: str, **kwargs: dict) -> Tuple[IFilter, Any, Context]:
-
-    if len(args) == 3:
-        return cast(IFilter, args[0]), args[1], cast(Context, args[2])
-
-    _self = None
-    _data = None
-    _context = None
-    _args: list = list(args)
-
-    if _args and isinstance(args[0], IFilter):
-        _self = cast(IFilter, _args[0])
-        _args.pop(0)
-
-    if _args:
-        _data = _args[0]
-        _args.pop(0)
-
-    if kwargs:
-        _data = _data or kwargs.get("data")
-        _context = _context or kwargs.get("context")
-
-    return cast(IFilter, _self), _data, cast(Context, _context) if _context else Context()
-
-
-def raise_events(func: Callable) -> Callable:
-    """
-    Decorates a filter method to raise events
-    """
-    @wraps(func)
-    def wrapper(*args: Any, **kwargs: Any) -> Callable:
-        filter, data, context = _parse_args(*args, **kwargs)
-
-        evt_args = FilterEventArgs(data, context, State.RUNNING)
-
-        filter._on_start(evt_args)
-
-        if evt_args.action in (Action.EXIT, Action.SKIP):
-            evt_args.state = State.SKIPPING if evt_args.action == Action.SKIP else State.EXITING
-            filter._on_exit(evt_args)
-            return evt_args.data
-
-        data = func(*args, **kwargs)
-
-        evt_args = FilterEventArgs(data, context, State.COMPLETING)
-        filter._on_end(evt_args)
-
-        if evt_args.action is Action.EXIT:
-            evt_args.state = State.EXITING
-            filter._on_exit(evt_args)
-            return evt_args.data
-
-        return data
-    return wrapper